refactor(session): extract session manager
